Finish refactoring tests
