+++ jmetal/core/test/test_solution.py
