+++ jmetal/runner/singleobjective/generationalGeneticAlgorithmRunnerWithObserver.py
@@ -4,7 +4,7 @@ from jmetal.operator.crossover import SBX
 from jmetal.operator.mutation import Polynomial
 from jmetal.operator.selection import BinaryTournament
 from jmetal.problem.singleobjectiveproblem import Sphere
-from jmetal.util.observer.observer import Observer
+from jmetal.util.observable import Observer
 
 
 def main():
