+++ jmetal/util/test/test_observable.py
